Backend/requirements.txt
asgiref==3.8.1
async-timeout==4.0.3
cachetools==4.2.4
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.4
django-cors-headers==4.3.1
django-filter==24.2
django-redis==5.4.0
djangorestframework==3.15.1
google-api-core==1.22.2
google-api-python-client==2.0.2
google-auth==1.21.2
google-auth-httplib2==0.0.4
googleapis-common-protos==1.52.0
httplib2==0.22.0
idna==3.7
mysqlclient==2.2.4
protobuf==5.26.1
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyparsing==3.1.2
pytz==2024.1
redis==5.0.3
requests==2.31.0
rsa==4.9
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.11.0
uritemplate==3.0.1
urllib3==2.2.1
